ExamPreparationChatbot(project)/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.12.0